Seperating Json Tests from GistTests
