+++ UnitTests/GistTests.cs
@@ -1,33 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using FluentAssertions;
+﻿using FluentAssertions;
 using IronGitHub.Entities;
 using NUnit.Framework;
-using ServiceStack.Text;
-using JsonExtensions = IronGitHub.JsonExtensions;
 
 namespace UnitTests
 {
-    [TestFixture]
-    class GistTests
-    {
-        [Test]
-        public void DictionaryShouldSerializeNullEntry()
-        {
-            var d = new Dictionary<string, Gist.NewGistPost.NewGistFile>();
-            d.Add("foo", new Gist.NewGistPost.NewGistFile());
-            d.Add("bar", null);
-            JsonExtensions.Init();
-
-            var json = JsonSerializer.SerializeToString(d);
-
-            json.Should().Contain("bar");
-        }
-    }
-
     [TestFixture]
     public class GistUrlParseTests
     {
